Changing folder structure Renaming az_iot_common to az_iot_core
