+++ CMakeLists.txt
@@ -44,6 +44,8 @@ endif()
 # SDK Clients and tests
 add_subdirectory(sdk/storage/blobs)
 add_subdirectory(sdk/iot/core)
+add_subdirectory(sdk/iot/hub)
+add_subdirectory(sdk/iot/provisioning)
 
 # PAL
 add_subdirectory(sdk/platform/noplatform)
@@ -62,7 +64,8 @@ endif()
 if (UNIT_TESTING)
   add_subdirectory(sdk/core/core/test/cmocka)
   add_subdirectory(sdk/storage/blobs/test/cmocka)
-  add_subdirectory(sdk/iot/core/tests/cmocka)
+  add_subdirectory(sdk/iot/hub/tests/cmocka)
+  add_subdirectory(sdk/iot/provisioning/tests/cmocka)
 
   if(NOT DEFINED ENV{AZ_SDK_C_NO_SAMPLES})
     add_subdirectory(sdk/samples/keyvault/keyvault/test/cmocka)
