+++ sdk/iot/core/tests/cmocka/test_az_iot_core.h
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+// Placeholder for int test_iot_common_(); declarations
