+++ sdk/iot/doc/mqtt_state_machine.md
