+++ sdk/iot/doc/resources/iot_provisioning_flow.puml
