+++ sdk/iot/hub/inc/az_iot_sas_token.h
