+++ sdk/iot/hub/src/az_iot_hub_client.c
