+++ sdk/iot/hub/tests/cmocka/CMakeLists.txt
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: MIT
 
 cmake_minimum_required (VERSION 3.10)
-set(TARGET_NAME "az_iot_core_test")
+set(TARGET_NAME "az_iot_hub_client_test")
 
 project (${TARGET_NAME} LANGUAGES C)
 
@@ -12,8 +12,12 @@ include(AddTestCMocka)
 
 add_cmocka_test(${TARGET_NAME} SOURCES
                 main.c
+                test_az_iot_hub_client_sas.c
+                test_az_iot_hub_client_telemetry.c
+                test_az_iot_hub_client.c
                 COMPILE_OPTIONS ${DEFAULT_C_COMPILE_FLAGS}
                 LINK_TARGETS
                     az_core
                     az_iot_core
+                    az_iot_hub_client
                 )
