+++ sdk/iot/provisioning/tests/cmocka/test_az_iot_provisioning_client.h
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+// Placeholder for int test_iot_provisioning_(); declarations
